[refactor]: move FA default custom-op registration to attention/utils/
Addresses the S2 finding from the post-merge review on #1373: per FastVideo
precedent (fastvideo/attention/utils/flash_attn_cute.py — the FP4 cute
custom-op wrapper #1388 says it mirrors) and vLLM's similar pattern
(vllm/utils/flashinfer.py), per-kernel custom-op wrappers live in
attention/utils/, not inline in the backend dispatcher.

Move the entire `if fa_version == "2": ... elif "3": ... elif "4": ... else:`
block — including the FA version probe, custom_op + register_fake +
setup_context + register_autograd, and `flash_attn_func_compilable` —
from `attention/backends/flash_attn.py` to a new sibling at
`attention/utils/flash_attn_default.py`. The backend now imports
`fa_version` and `flash_attn_func_compilable` from there; the rest of the
backend (FlashAttentionBackend, FlashAttentionImpl, FP4 quantize helpers,
masked-branch routing) is unchanged.

Benefits:
  - Internal consistency with flash_attn_cute.py (the FP4 template the
    file-level comment claims to mirror) and flash_attn_no_pad.py (the
    masked/varlen wrappers, already in utils/).
  - A future backend that wants the same traceable FA2/FA3/FA4 op can
    import from utils/ without depending on the FA backend's dispatcher.
  - Cleaner separation between backend registry/dispatch logic and
    per-kernel custom-op plumbing.

No semantic change. Verified there is exactly one registration of
`fastvideo::_flash_attn_default_forward` after the move (grep clean).
Update `test_flash_attn_default_custom_op.py` to import from the utils
module (the test referenced `fa_backend._fa_default`, a name that lived
inside the moved block).
